Convert label updates to a handler and remove all traces of pre-handler test setup.
